% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{saveObject,sf-method}
\alias{saveObject,sf-method}
\title{Save \code{sf} to disk for \code{alabaster}}
\usage{
\S4method{saveObject}{sf}(x, path)
}
\arguments{
\item{x}{A Bioconductor object of the specified class.}

\item{path}{String containing the path to a directory in which to save \code{x}.}
}
\description{
\code{sf} data frames are saved as GeoParquet.
}
\examples{
# example code

}
