% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImages.R
\name{readBioFormatsImage}
\alias{readBioFormatsImage}
\title{Read \code{BioFormatsImage} from \code{alabaster} on disk representation}
\usage{
readBioFormatsImage(path, metadata = NULL, ...)
}
\arguments{
\item{path}{String containing a path to a directory, itself created with a \code{\link[alabaster.base]{saveObject}} method.}

\item{metadata}{Named list containing metadata for the object - most importantly, the \code{type} field that controls dispatch to the correct loading function.
If \code{NULL}, this is automatically read by \code{\link[alabaster.base]{readObjectFile}(path)}.}

\item{...}{Ignored, but used for other methods.}
}
\value{
A \code{\link{BioFormatsImage}} object for SFE.
}
\description{
Reads the metadata and \code{imgSource} will point to the file within the
on disk representation of SFE. The image itself should not be moved or the
\code{BioFormatsImabe} object will no longer work.
}
\examples{
# example code

}
\seealso{
Other readObject-SFE-image: 
\code{\link{readExtImage}()},
\code{\link{readSpatRaster}()}
}
\concept{readObject-SFE-image}
