% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSFE.R
\name{saveObject,SpatialFeatureExperiment-method}
\alias{saveObject,SpatialFeatureExperiment-method}
\title{Save a SpatialFeatureExperiment object}
\usage{
\S4method{saveObject}{SpatialFeatureExperiment}(x, path, ...)
}
\arguments{
\item{x}{A Bioconductor object of the specified class.}

\item{path}{String containing the path to a directory in which to save \code{x}.}

\item{...}{Additional named arguments to pass to specific methods.}
}
\description{
Save SFE objects to disk in an interoperable, language agnostic format that
may also facilitate out of memory operations via HDF5 (non-spatial, inherited
from \code{alabaster.sce}) and Apache Parquet (geometries).
}
\details{
There's no new arguments for \code{...} for the SFE,
\code{SpatialExperiment}, and \code{SingleCellExperiment} methods, but there
is an argument that can be specified for \code{...} in the
\code{SummarizedExperiment} method of \code{saveObject}.

At present, spatial results in \code{featureData} for geometries and
dimension reductions (see \code{\link{colFeatureData}}) and parameters of
spatial analyses (see \code{\link{getParams}}) are not save because those
parts of the SFE object are more experimental and are subject to change.
\code{colFeatureData} where global spatial results are stored for columns of
\code{colData(x)} is saved by the \code{alabaster} \code{saveObject} method
for \code{DataFrame}.
}
